backend/requirements.txt
alembic==1.14.0
amqp==5.2.0
aniso8601==9.0.1
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.2.0
blinker==1.8.2
cachelib==0.9.0
celery==5.4.0
certifi==2021.5.30
chardet>=3.0.2,<5
checksumdir==1.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.1.1
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
executing==2.1.0
fastjsonschema==2.21.0
filelock==3.13.1
Flask==3.0.3
Flask-Caching==2.3.0
flask-cors==5.0.1
Flask-Excel==0.0.7
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-MySQL==1.5.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.5.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
gunicorn==20.1.0
idna==2.10
importlib_resources==6.4.5
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kombu==5.4.2
lml==0.1.0
Mako==1.3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.26.4
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
peewee==3.17.8
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
Pygments==2.18.0
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
pytz==2024.2
# pywin32==308
pyzmq==26.2.0
redis==5.2.0
referencing==0.35.1
requests==2.25.1
rope==0.19.0
rpds-py==0.21.0
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.36
sqlite-web==0.6.4
sqlparse==0.5.1
stack-data==0.6.3
texttable==1.7.0
timeago==1.0.15
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.5
vine==5.1.0
virtualenv==20.25.1
vulture==2.13
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
WTForms==3.1.2
yarg==0.1.9
